+++ massiv/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Deprecated `indexWith` in favor of `indexAssert`
 * Addition of scans: `sscanl`, `sscanl1`, `sprescanl`, `spostscanl` and `spostscanlAcc`
+* Expose `unsafePrefIndex`
 
 # 1.0.2